requirements.txt
numpy
opencv-python
pillow
scikit-learn
scipy
torch
torchvision
tqdm
